Refactor `in` and `check` to have separate main
- to be consistent with `out`
