+++ ci/test-and-build.sh
@@ -11,8 +11,8 @@ export INPUTDIR=$PWD/go
 pushd go/src/github.com/pivotal-cf/email-resource
   ginkgo -r -p -race "$@"
 
-  go build -tags netgo -a -o bin/check ./actions/check
-  go build -tags netgo -a -o bin/in ./actions/in
+  go build -tags netgo -a -o bin/check ./cmds/check
+  go build -tags netgo -a -o bin/in ./cmds/in
   go build -tags netgo -a -o bin/out ./cmds/out
 popd
