+++ cmds/check/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/pivotal-cf/email-resource/actions/check"
+)
+
+func main() {
+	check.Run()
+}
